fix: fix hero
